requirements.txt
aioredis==2.0.1
aiohttp==3.8.4
python-dotenv==1.0.0
